Moved TCs from removeGlobalStorage_webhook_test.go to respective *_webhook_test.go
